add support for kwargs for loss and maximizers in tests
