+++ src/regularized/soft_argmax.jl
@@ -7,7 +7,7 @@ Corresponds to regularized prediction on the probability simplex with entropic p
 """
 struct SoftArgmax <: AbstractRegularized end
 
-(::SoftArgmax)(z) = soft_argmax(z)
+(::SoftArgmax)(z; kwargs...) = soft_argmax(z)
 compute_regularization(::SoftArgmax, y) = soft_argmax_regularization(y)
 
 function soft_argmax(z::AbstractVector)
