+++ test/InferOptTestUtils/InferOptTestUtils.jl
@@ -1,7 +1,7 @@
 module InferOptTestUtils
 
 using Flux
-using Flux.Losses
+using Flux.Losses: mse
 using Graphs
 using GridGraphs
 using LinearAlgebra
@@ -18,7 +18,7 @@ include("loss.jl")
 include("pipeline.jl")
 
 export DECREASE, EPOCHS, NB_FEATURES, NB_INSTANCES, NOISE_STD, VERBOSE
-export shortest_path_maximizer
+export shortest_path_maximizer, max_pricing, g, h
 export encoder_factory, generate_dataset
 export mse, mape, normalized_mape, hamming_distance, normalized_hamming_distance
 export init_perf, update_perf!
@@ -27,5 +27,6 @@ export PipelineLossImitation, PipelineLossImitationθ, PipelineLossImitationθy
 export PipelineLossExperience
 export PipelineLossImitationLoss
 export test_pipeline!
+export mse_kw, identity_kw
 
 end
