+++ test/generalized_maximizer.jl
@@ -1,24 +1,7 @@
-# TODO: update to new setup
+@testitem "Generalized maximizer basics" default_imports = false begin
+    include("InferOptTestUtils/InferOptTestUtils.jl")
+    using InferOpt, .InferOptTestUtils, Test
 
-using Random
-using LinearAlgebra
-
-Random.seed!(67)
-
-verbose = false
-
-CC = 10
-function max_pricing(θ::AbstractVector; instance::AbstractMatrix)
-    @assert length(θ) == size(instance, 1)
-    @assert length(θ) == size(instance, 2)
-    weights = θ .- instance
-    return weights .>= 0
-end
-
-g(y; kwargs...) = vec(sum(y; dims=2))
-h(y; instance) = -sum(dij * yij for (dij, yij) in zip(instance, y))
-
-@testset "Generalized maximizer basics" begin
     instance = [
         1.0 2.0 0.0
         1.0 0.0 1.0
@@ -39,205 +22,76 @@ h(y; instance) = -sum(dij * yij for (dij, yij) in zip(instance, y))
     @test val == θ' * g(y) + h(y; instance)
 end
 
-nb_features = 5
-encoder_factory() = Chain(Dense(nb_features => 1; bias=false), dropfirstdim)#, make_positive)
-true_encoder = encoder_factory()
-generalized_maximizer = GeneralizedMaximizer(max_pricing, g, h)
+@testitem "Generalized maximizer - imit - MSE PerturbedAdditive" default_imports = false begin
+    include("InferOptTestUtils/InferOptTestUtils.jl")
+    using InferOpt, .InferOptTestUtils, Random
+    Random.seed!(63)
 
-data_train, data_test = generate_dataset(
-    true_encoder,
-    generalized_maximizer;
-    nb_features=nb_features,
-    instance_dim=5,
-    nb_instances=100,
-    noise_std=0.0,
-);
+    true_encoder = encoder_factory()
 
-function cost(y; instance)
-    return -objective_value(generalized_maximizer, true_encoder(instance), y; instance)
-end
-error_function(ŷ, y) = hamming_distance(ŷ, y)
+    generalized_maximizer = GeneralizedMaximizer(max_pricing, g, h)
+    function cost(y; instance)
+        return -objective_value(generalized_maximizer, true_encoder(instance), y; instance)
+    end
 
-mse_loss(y1, y2; kwargs...) = Flux.Losses.mse(y1, y2)
-identity_maximizer(θ; kwargs...) = identity(θ)
-
-pipelines_imitation_y = [
-    # Interpolation
-    # (
-    #     encoder=encoder_factory(),
-    #     maximizer=Interpolation(generalized_maximizer; λ=5.0),
-    #     loss=Flux.Losses.mse,
-    # ),
-    # Perturbed + FYL
-    (
-        encoder=encoder_factory(),
-        maximizer=identity_maximizer,
-        loss=FenchelYoungLoss(
-            PerturbedAdditive(generalized_maximizer; ε=1.0, nb_samples=5)
-        ),
-    ),
-    (
-        encoder=encoder_factory(),
-        maximizer=identity_maximizer,
-        loss=FenchelYoungLoss(
-            PerturbedMultiplicative(generalized_maximizer; ε=1.0, nb_samples=5)
-        ),
-    ),
-    # Perturbed + other loss
-    (
-        encoder=encoder_factory(),
+    test_pipeline!(
+        PipelineLossImitation;
+        instance_dim=5,
+        true_maximizer=max_pricing,
         maximizer=PerturbedAdditive(generalized_maximizer; ε=1.0, nb_samples=10),
-        loss=mse_loss,
-    ),
-    (
-        encoder=encoder_factory(),
-        maximizer=PerturbedMultiplicative(generalized_maximizer; ε=1.0, nb_samples=50), # more samples were needed here
-        loss=mse_loss,
-    ),
-    # # Generic regularized + FYL
-    # (
-    #     encoder=encoder_factory(),
-    #     maximizer=identity,
-    #     loss=FenchelYoungLoss(
-    #         RegularizedGeneric(generalized_maximizer, half_square_norm, identity)
-    #     ),
-    # ),
-    # # Generic regularized + other loss
-    # (
-    #     encoder=encoder_factory(),
-    #     maximizer=RegularizedGeneric(generalized_maximizer, half_square_norm, identity),
-    #     loss=Flux.Losses.mse,
-    # ),
-]
-
-pipelines_imitation_θ = [
-# SPO+
-(
-    encoder=encoder_factory(),
-    maximizer=identity_maximizer,
-    loss=SPOPlusLoss(generalized_maximizer),
-)]
-
-pipelines_experience = [
-    (
-        encoder=encoder_factory(),
-        maximizer=identity_maximizer,
-        loss=Pushforward(
-            PerturbedAdditive(generalized_maximizer; ε=1.0, nb_samples=50), cost
-        ),
-    ),
-    (
-        encoder=encoder_factory(),
-        maximizer=identity_maximizer,
-        loss=Pushforward(
-            PerturbedMultiplicative(generalized_maximizer; ε=1.0, nb_samples=50), cost
-        ),
-    ),
-    (
-        encoder=encoder_factory(),
-        maximizer=identity_maximizer,
-        loss=Pushforward(
-            PerturbedAdditive(
-                generalized_maximizer; ε=1.0, nb_samples=50, is_parallel=true
-            ),
-            cost,
-        ),
-    ),
-    (
-        encoder=encoder_factory(),
-        maximizer=identity_maximizer,
-        loss=Pushforward(
-            PerturbedMultiplicative(
-                generalized_maximizer; ε=1.0, nb_samples=50, is_parallel=true
-            ),
-            cost,
-        ),
-    ),
-    # (
-    #     encoder=encoder_factory(),
-    #     maximizer=identity,
-    #     loss=Pushforward(
-    #         RegularizedGeneric(true_maximizer, half_square_norm, identity), cost
-    #     ),
-    # ),
-]
-
-for pipeline in pipelines_imitation_y
-    pipeline = deepcopy(pipeline)
-    (; encoder, maximizer, loss) = pipeline
-    function pipeline_loss_imitation_y(x, θ, y)
-        return loss(maximizer(encoder(x); instance=x), y; instance=x)
-    end
-    test_pipeline!(
-        pipeline,
-        pipeline_loss_imitation_y;
-        true_encoder=true_encoder,
-        true_maximizer=generalized_maximizer,
-        data_train=data_train,
-        data_test=data_test,
-        error_function=error_function,
-        cost=cost,
-        epochs=1000,
-        verbose=verbose,
-        setting_name="generalized maximizer - imitation_y",
+        loss=mse_kw,
+        error_function=hamming_distance,
+        cost,
+        true_encoder,
     )
 end
 
-for pipeline in pipelines_imitation_θ
-    pipeline_1 = deepcopy(pipeline)
-    (; encoder, maximizer, loss) = pipeline_1
-    function pipeline_loss_imitation_θ(x, θ, y)
-        return loss(maximizer(encoder(x); instance=x), θ; instance=x)
-    end
-    test_pipeline!(
-        pipeline_1,
-        pipeline_loss_imitation_θ;
-        true_encoder=true_encoder,
-        true_maximizer=generalized_maximizer,
-        data_train=data_train,
-        data_test=data_test,
-        error_function=error_function,
-        cost=cost,
-        epochs=100,
-        verbose=verbose,
-        setting_name="generalized maximizer - imitation_θ",
-    )
+@testitem "Generalized maximizer - imit - MSE PerturbedMultiplicative" default_imports =
+    false begin
+    include("InferOptTestUtils/InferOptTestUtils.jl")
+    using InferOpt, .InferOptTestUtils, Random
+    Random.seed!(63)
 
-    pipeline_2 = deepcopy(pipeline)
-    (; encoder, maximizer, loss) = pipeline_2
-    function pipeline_loss_imitation_θ(x, θ, y)
-        return loss(maximizer(encoder(x); instance=x), θ, y; instance=x)
+    true_encoder = encoder_factory()
+
+    generalized_maximizer = GeneralizedMaximizer(max_pricing, g, h)
+    function cost(y; instance)
+        return -objective_value(generalized_maximizer, true_encoder(instance), y; instance)
     end
+
     test_pipeline!(
-        pipeline_2,
-        pipeline_loss_imitation_θ;
-        true_encoder=true_encoder,
-        true_maximizer=generalized_maximizer,
-        data_train=data_train,
-        data_test=data_test,
-        error_function=error_function,
-        cost=cost,
-        epochs=100,
-        verbose=verbose,
-        setting_name="generalized maximizer - imitation_θ - precomputed y_true",
+        PipelineLossImitation;
+        instance_dim=5,
+        true_maximizer=max_pricing,
+        maximizer=PerturbedMultiplicative(generalized_maximizer; ε=1.0, nb_samples=10),
+        loss=mse_kw,
+        error_function=hamming_distance,
+        cost,
+        true_encoder,
     )
 end
 
-for pipeline in pipelines_experience
-    pipeline = deepcopy(pipeline)
-    (; encoder, maximizer, loss) = pipeline
-    pipeline_loss_experience(x, θ, y) = loss(maximizer(encoder(x)); instance=x)
-    test_pipeline!(
-        pipeline,
-        pipeline_loss_experience;
-        true_encoder=true_encoder,
-        true_maximizer=generalized_maximizer,
-        data_train=data_train,
-        data_test=data_test,
-        error_function=error_function,
-        cost=cost,
-        epochs=1000,
-        verbose=verbose,
-        setting_name="generalized maximizer - experience",
-    )
-end
+# @testitem "Generalized maximizer - imit - FYL PerturbedAdditive" default_imports = false begin
+#     include("InferOptTestUtils/InferOptTestUtils.jl")
+#     using InferOpt, .InferOptTestUtils, Random
+#     Random.seed!(63)
+
+#     true_encoder = encoder_factory()
+
+#     generalized_maximizer = GeneralizedMaximizer(max_pricing, g, h)
+#     function cost(y; instance)
+#         return -objective_value(generalized_maximizer, true_encoder(instance), y; instance)
+#     end
+
+#     test_pipeline!(
+#         PipelineLossImitation;
+#         instance_dim=5,
+#         true_maximizer=max_pricing,
+#         maximizer=identity_kw,
+#         loss=FenchelYoungLoss(
+#             PerturbedAdditive(generalized_maximizer; ε=1.0, nb_samples=5)
+#         ),
+#         error_function=hamming_distance,
+#         true_encoder,
+#     )
+# end
